Many updates, see details below
Moved utility functions to it's own file (because they shouldn't be in the main driver): doomUtils.py

Moved feature extractor function to it's own file (because they shouldn't be in qlearning agent): extractor.py Note: because of this, we only need one learningAgent file. The only difference between the old qlearningAgent.py and healthlearningAgent.py was the feature extractor.

state variable should now include the scenario, which should be a string. This is used in the getFeatures function in extractor.py to dispatch to the correct function.

health.py now works, although not as good as basic.py.